Restructure python namespaces, and add search to events.get.
This is a pretty huge commit and it may have some bugs.  Python
filenames now make more sense, and all helper functions have a logical
place.  Also added searching and sorting to results from an event query.
Starred events are also guranteed to show up on top, even if they are
far in the future.